Moved classes around to seperate files for clarity
